checkpoint before migrate
