+++ src-tauri/Cargo.toml
@@ -24,10 +24,13 @@ tauri = { version = "1.6.0", features = [
     "window-minimize",
     "window-start-dragging",
 ] }
-window-vibrancy = "0.4.0"
 
 [features]
 # this feature is used for production builds or when `devPath` points to the filesystem and the built-in dev server is disabled.
 # If you use cargo directly instead of tauri's cli you can use this feature flag to switch between tauri's `dev` and `build` modes.
 # DO NOT REMOVE!!
 custom-protocol = ["tauri/custom-protocol"]
+
+[lib]
+name = "app_lib"
+crate-type = ["staticlib", "cdylib", "rlib"]
