+++ src-tauri/tauri.conf.json
@@ -11,43 +11,7 @@
   },
   "tauri": {
     "allowlist": {
-      "all": false,
-      "window": {
-        "all": false,
-        "center": false,
-        "close": true,
-        "create": false,
-        "hide": false,
-        "maximize": true,
-        "minimize": true,
-        "print": false,
-        "requestUserAttention": false,
-        "setAlwaysOnTop": false,
-        "setClosable": false,
-        "setContentProtected": false,
-        "setCursorGrab": false,
-        "setCursorIcon": false,
-        "setCursorPosition": false,
-        "setCursorVisible": false,
-        "setDecorations": false,
-        "setFocus": false,
-        "setFullscreen": false,
-        "setIcon": false,
-        "setIgnoreCursorEvents": false,
-        "setMaxSize": false,
-        "setMaximizable": false,
-        "setMinSize": false,
-        "setMinimizable": false,
-        "setPosition": false,
-        "setResizable": false,
-        "setSize": false,
-        "setSkipTaskbar": false,
-        "setTitle": false,
-        "show": false,
-        "startDragging": true,
-        "unmaximize": true,
-        "unminimize": false
-      }
+      "all": false
     },
     "bundle": {
       "active": true,
@@ -94,9 +58,7 @@
         "height": 600,
         "resizable": true,
         "title": "My app",
-        "width": 800,
-        "transparent": true,
-        "decorations": false
+        "width": 800
       }
     ]
   }
